@@ -1,7 +1,4 @@
-﻿using Dots.Controls;
-using Microsoft.AppCenter.Crashes;
-using Mopups.Services;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -14,9 +11,10 @@ namespace Dots.Helpers
     {
         public async Task ShowPopup(Exception ex)
         {
+            /*
             Crashes.TrackError(ex);
             await MopupService.Instance.PushAsync(new ErrorPopup(ex.Message, ex.StackTrace));
-
+            */
         }
     }
 }
