@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Avalonia.Data.Converters;
+using System;
 using System.Globalization;
 
 namespace Dots.Helpers
@@ -14,6 +15,19 @@ namespace Dots.Helpers
                 else return "🐧";
             }
             return null;
+        }
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    public class StringIsNullOrEmptyConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is string str && !string.IsNullOrEmpty(str)) return true;
+            else return false;
         }
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
